refactor(rust): make secure channel structure more consistent
